requirements.txt
asgiref==3.7.2
boto3==1.20.26
botocore==1.23.54
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cryptography==41.0.7
diff-match-patch==20230430
distlib==0.3.4
dj-database-url==2.1.0
Django==4.2.7
django-anymail==10.2
django-cors-headers==3.14.0
django-import-export==4.1.1
django-jazzmin==2.6.0
django-storages==1.12.3
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
drf-yasg==1.21.7
environs==10.0.0
filelock==3.7.1
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
jmespath==0.10.0
marshmallow==3.20.1
mysqlclient==2.2.0
num2words==0.5.13
packaging==23.2
pillow==10.4.0
pipenv==2022.7.4
platformdirs==2.5.2
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
reportlab==4.2.2
requests==2.31.0
s3transfer==0.5.2
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
stripe==7.9.0
tablib==3.5.0
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.18
virtualenv==20.16.2
virtualenv-clone==0.5.7
whitenoise==6.7.0
